requirements.txt
blinker==1.8.2
click==8.1.8
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
importlib_metadata==8.5.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
packaging==25.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.3.5
pytest-mock==3.14.0
python-dotenv
SQLAlchemy==2.0.40
tomli==2.2.1
typing_extensions==4.13.2
Werkzeug==3.0.6
zipp==3.20.2
